Amending previous commit, missing Material Symbols Outlined font
On the previous commit where we optimised loading of critical JS/Styles, we created a new Google Fonts link to import the fonts used on the project and forgot to add the Material Symbols Outlined font back before submitting the changes.

Adding it back has not affected the website performance.